add: basic styles to blog section, changed and fixed some component's styles.
