+++ src/components/ComponentButtons/buttons.scss
@@ -18,7 +18,6 @@
   box-shadow: 0 5px 10px rgba(0, 0, 0, 0.35);
   color: $purple;
   font-size: 20px;
-  height: 60px;
   letter-spacing: 0.2em;
   transition: 0.5s;
 }
