+++ src/components/nav-menu/nav-menu.js
@@ -1,16 +1,21 @@
 import React from "react"
-import "./nav-menu.scss"
 import { Link } from "gatsby"
+import "./nav-menu.scss"
 
 const Menu = () => (
   <React.Fragment>
-    <div className="nav__menu">
-      <Link to="" className="nav__link">
-        iniciar sesión
+    <div className="nav__menu section__wrapper">
+      <Link to="">
+        <img src="atana-logo.svg" alt="átana" className="header__logo" />
       </Link>
-      <Link to="" className="nav__btn btn btn--primary">
-        Suscribirse
-      </Link>
+      <div>
+        <Link to="" className="nav__link">
+          iniciar sesión
+        </Link>
+        <Link to="" className="nav__btn btn btn--primary">
+          Suscribirse
+        </Link>
+      </div>
     </div>
   </React.Fragment>
 )
