+++ src/components/nav-menu/nav-menu.scss
@@ -3,6 +3,7 @@
 .nav__btn {
   font-size: 16px;
   height: 30px;
+  margin: 0 0 0 15px;
   padding: 6px;
   text-decoration: none;
 
@@ -16,7 +17,6 @@
   border-bottom: 1px solid $white;
   color: $white;
   font-weight: bold;
-  margin-right: 21px;
   text-align: center;
   text-decoration: none;
   text-transform: capitalize;
@@ -25,20 +25,14 @@
 .nav__menu {
   align-items: center;
   display: flex;
-  flex-direction: row;
+  flex-direction: column-reverse;
   justify-content: center;
-  left: 0;
-  margin: 0 20px;
-  padding-top: 30px;
-  position: absolute;
-  right: 0;
-  top: 0;
+  padding: 10px 0;
 
   @media screen and (min-width: 768px) {
     flex-direction: row;
-    justify-content: flex-end;
+    justify-content: space-between;
     margin-bottom: 40px;
-    margin-top: -83px;
     position: static;
     right: 442px;
   }
