+++ src/pages/blog.js
@@ -1,5 +1,9 @@
 import React from "react"
 import { Link, graphql } from "gatsby"
+import SEO from "../components/seo"
+import Menu from "../components/nav-menu/nav-menu"
+import Footer from "../components/footer/footer"
+import "../styles/blog/section-blog.scss"
 
 class BlogIndex extends React.Component {
   render() {
@@ -10,17 +14,32 @@ class BlogIndex extends React.Component {
       const title = node.frontmatter.title || node.fields.slug
 
       return (
-        <div key={node.fields.slug}>
-          <Link style={{ boxShadow: `none` }} to={node.fields.slug}>
-            {title}
-          </Link>
-          <small>{node.frontmatter.date}</small>
-          <p
-            dangerouslySetInnerHTML={{
-              __html: node.frontmatter.description || node.excerpt,
-            }}
-          />
-        </div>
+        <React.Fragment>
+          <SEO title="Home" />
+          <div className="menu">
+            <Menu />
+          </div>
+          <div className="section__wrapper">
+            <h2 className="blog__title">
+              Átana blog
+            </h2>
+            <div key={node.fields.slug} className="post__wrapper">
+              <Link to={node.fields.slug} className="post">
+                <h4 className="post__title">
+                  {title}
+                </h4>
+                <small className="post__date post__preview">{node.frontmatter.date}</small>
+                <p
+                  dangerouslySetInnerHTML={{
+                    __html: node.frontmatter.description || node.excerpt,
+                  }}
+                  className="post__preview"
+                />
+              </Link>
+            </div>
+          </div>
+          <Footer />
+        </React.Fragment>
       )
     })
   }
