+++ src/sections/Header/header.scss
@@ -54,13 +54,13 @@
 .header__logo {
   display: block;
   margin: 0 auto;
-  padding-top: 120px;
+  padding: 50px 0 30px;
   width: 73%;
 
   @media screen and (min-width: 768px) {
     margin: 0;
     padding-top: 30px;
-    width: 12%;
+    width: 105px;
   }
 }
