+++ src/sections/footer-section/footer.js
@@ -1,29 +1,17 @@
 import React from "react"
 import Greetings from "../../components/greetings/greetings"
-import { Link } from "gatsby"
+import Footer from "../../components/footer/footer"
 import "./footer.scss"
 
-const Footer = () => {
+const HomeFooter = () => {
   return (
     <section className="contact">
       <div className="section__wrapper">
         <Greetings />
       </div>
-      <div className="footer__gradient">
-        <div className="contact__social">
-          <Link to="">
-            <img src="atana-logo.svg" alt="Átana" className="footer__logo" />
-          </Link>
-          <div className="social__container">
-            <p>Síguenos en nuestras redes sociales:</p>
-            <Link to="" className="social social--facebook" />
-            <Link to="" className="social social--instagram" />
-            <Link to="" className="social social--whatsapp" />
-          </div>
-        </div>
-      </div>
+    <Footer />
     </section>
   )
 }
 
-export default Footer
+export default HomeFooter
