+++ src/sections/footer-section/footer.scss
@@ -27,8 +27,9 @@
 }
 
 .social__container {
-  width: 50%;
+  color: white;
   text-align: center;
+  width: 50%;
   
   @media screen and (min-width: 400px) {
     width: 50%;
@@ -44,7 +45,7 @@
   background-repeat: no-repeat;
   background-size: cover;
   height: 200px;
-  padding: 50px 0;
+  padding: 60px 0;
   position: relative;
   width: 100%;
 }
