+++ src/styles/blog/section-blog.scss
@@ -1,0 +1,42 @@
+.nav__bar {
+  align-items: center;
+  background: linear-gradient(90deg, #8B1C63 0%, #BD385B 97.4%);
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
+
+.nav__logo {
+  width: 105px;
+}
+
+.post {
+  text-decoration: none;
+}
+
+.post__date {
+  font-size: 14px;
+}
+
+.post__title {
+  font-size: 24px;
+  font-weight: 600;
+  letter-spacing: 0.08px;
+}
+
+.post__preview {
+  color: #484848;
+}
+
+.post__wrapper {
+  border-bottom: 1px solid #D5D5D5;
+  padding: 50px 0 20px;
+}
+
+.blog__title {
+  padding: 70px 0 30px;
+}
+
+.menu {
+  background: linear-gradient(90deg, #8B1C63 0%, #BD385B 97.4%);
+}
